backend/requirements.txt
Django
djangorestframework
django-cors-headers
celery
redis
psycopg2-binary
langgraph
langchain
langchain_openai
langchain_google_genai
langchain_deepseek
python-dotenv
cryptography
gunicorn
uvicorn
django-celery-results
channels
daphne
google-genai
langchain-ollama
